rename Files.tsx to FileHandler.tsx
